requirements.txt
streamlit==1.31.1
python-dotenv==1.0.0
pymongo==4.6.0
openai==1.2.4
streamlit-authenticator==0.2.3
st_clickable_images==0.0.3
python-dotenv==1.0.0
authlib==1.3.0
secrets==1.0.2